Added sliders for constant changing
Constants were just being modified within the particle class. This did not effect the drawing as that happens within the ParticleSystem class. If the slider is changed a pointer to the particleSystem is then called that calls the relevant function. This then goes through a for loop for each of the particles updating the new constant value within them